requirements.txt
streamlit
pyzxing
pillow
isbnlib
pandas
pyheif